use_query function with replaced usages
