+++ ui/src/pages/campaign/campaign_page.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use graphql_client::GraphQLQuery;
 use serde_json::json;
 use wasm_bindgen_futures::spawn_local;
-use yew::{function_component, html, use_effect_with_deps, use_state, Callback, Properties};
+use yew::{function_component, html, Callback, Properties};
 use yew_router::history::History;
 use yew_router::hooks::use_history;
 
@@ -12,59 +12,13 @@ use crate::{
     navigation::AppRoute,
     pages::{
         campaign::initiative_tracker::{InitiativeInfo, InitiativeTracker},
-        character_sheet::sheet::{Character, CharacterList},
+        character_sheet::sheet::CharacterList,
     },
     services::{
-        self, characters_query, delete_campaign_mutation, CharactersQuery, DeleteCampaignMutation,
-        GraphQLResponse,
+        self, characters_query, delete_campaign_mutation, use_query::use_query, CharactersQuery,
+        DeleteCampaignMutation, GraphQLResponse,
     },
 };
-
-#[derive(Clone)]
-struct MyGraphQLRequest {
-    pub data: Option<Vec<Character>>,
-    pub error: Option<String>,
-}
-
-fn use_characters_query() -> MyGraphQLRequest {
-    let state = use_state(|| MyGraphQLRequest {
-        data: None,
-        error: None,
-    });
-
-    let effect_state = state.clone();
-
-    use_effect_with_deps(
-        move |_| {
-            spawn_local(async move {
-                let variables = characters_query::Variables {};
-                let request_body = CharactersQuery::build_query(variables);
-                let request_json = &json!(request_body);
-
-                let request = services::build_request(request_json).await;
-
-                if let Ok(response) = request {
-                    let json = response.json::<GraphQLResponse<CharacterList>>().await;
-                    match json {
-                        Ok(responser) => effect_state.set(MyGraphQLRequest {
-                            data: Some(responser.data.characters),
-                            error: None,
-                        }),
-                        Err(error) => effect_state.set(MyGraphQLRequest {
-                            data: None,
-                            error: Some(error.to_string()),
-                        }),
-                    }
-                }
-            });
-
-            || ()
-        },
-        (),
-    );
-
-    (*state).clone()
-}
 
 #[derive(Properties, Clone, PartialEq)]
 pub struct CampaignProps {
@@ -73,7 +27,10 @@ pub struct CampaignProps {
 
 #[function_component(CampaignPage)]
 pub fn campaign_page(props: &CampaignProps) -> Html {
-    let characters = use_characters_query();
+    let variables = characters_query::Variables {};
+    let request_body = CharactersQuery::build_query(variables);
+    let request_json = &json!(request_body);
+    let characters = use_query::<CharacterList>(request_json);
     let history = use_history().expect("history should be available");
 
     if characters.data.is_none() {
@@ -109,6 +66,7 @@ pub fn campaign_page(props: &CampaignProps) -> Html {
     let characters_initiative: HashMap<String, InitiativeInfo> = characters
         .data
         .unwrap_or_default()
+        .characters
         .iter()
         .map(|character| {
             (
