+++ ui/src/pages/campaigns.rs
@@ -2,7 +2,7 @@ use graphql_client::GraphQLQuery;
 use serde::Deserialize;
 use serde_json::json;
 use wasm_bindgen_futures::spawn_local;
-use yew::{function_component, html, use_effect_with_deps, use_state, Callback, Html};
+use yew::{function_component, html, use_state, Callback, Html};
 use yew_router::prelude::Link;
 
 use crate::{
@@ -11,6 +11,7 @@ use crate::{
     services::{
         self, campaigns_query,
         new_campaign_mutation::{self, NewCampaign},
+        use_query::use_query,
         CampaignsQuery, GraphQLResponse, NewCampaignMutation,
     },
 };
@@ -28,95 +29,12 @@ pub struct CampaignList {
     pub campaigns: Vec<Campaign>,
 }
 
-#[derive(Clone)]
-struct MyGraphQLRequest {
-    pub data: Option<Vec<Campaign>>,
-    pub error: Option<String>,
-}
-
-fn use_campaigns_query() -> MyGraphQLRequest {
-    let state = use_state(|| MyGraphQLRequest {
-        data: None,
-        error: None,
-    });
-
-    let effect_state = state.clone();
-
-    use_effect_with_deps(
-        move |_| {
-            spawn_local(async move {
-                let variables = campaigns_query::Variables {};
-                let request_body = CampaignsQuery::build_query(variables);
-                let request_json = &json!(request_body);
-
-                let request = services::build_request(request_json).await;
-                if let Ok(response) = request {
-                    let json = response.json::<GraphQLResponse<CampaignList>>().await;
-                    match json {
-                        Ok(responser) => effect_state.set(MyGraphQLRequest {
-                            data: Some(responser.data.campaigns),
-                            error: None,
-                        }),
-                        Err(error) => effect_state.set(MyGraphQLRequest {
-                            data: None,
-                            error: Some(error.to_string()),
-                        }),
-                    }
-                }
-            });
-
-            || ()
-        },
-        (),
-    );
-
-    (*state).clone()
-}
-
-// TODO: Do a more generalized query function
-// #[derive(Clone)]
-// struct MySmallerGraphQLRequest<T> {
-//     pub data: Option<T>,
-//     pub error: Option<String>,
-// }
-
-// fn use_smaller_query<'a, T>(request_json: &serde_json::Value) -> MySmallerGraphQLRequest<T>
-// where T: Deserialize<'a> + Clone {
-//     let state = use_state(|| MySmallerGraphQLRequest {
-//         data: None,
-//         error: None,
-//     });
-
-//     let effect_state = state.clone();
-
-//     let request_json = request_json.clone();
-//     use_effect_with_deps(move |_| {
-//         spawn_local(async move {
-//             let request = services::build_request(&request_json).await;
-//             if let Ok(response) = request {
-//                 let json = response.json::<GraphQLResponse<T>>().await;
-//                 match json {
-//                     Ok(responser) => effect_state.set(MySmallerGraphQLRequest {
-//                         data: Some(responser.data),
-//                         error: None,
-//                     }),
-//                     Err(error) => effect_state.set(MySmallerGraphQLRequest {
-//                         data: None,
-//                         error: Some(error.to_string()),
-//                     }),
-//                 }
-//             }
-//         });
-
-//         || ()
-//     }, ());
-
-//     (*state).clone()
-// }
-
 #[function_component(CampaignsPage)]
 pub fn campaigns_page() -> Html {
-    let query = use_campaigns_query();
+    let variables = campaigns_query::Variables {};
+    let request_body = CampaignsQuery::build_query(variables);
+    let request_json = &json!(request_body);
+    let query = use_query::<CampaignList>(request_json);
 
     let new_name = use_state(|| "".to_string());
     let new_description = use_state(|| "".to_string());
@@ -168,8 +86,8 @@ pub fn campaigns_page() -> Html {
         <section class="list-page">
             <>{query.error.unwrap_or_else(|| "".to_string())}</>
             {
-                if let Some(campaigns) = &query.data {
-                    campaigns.iter().map(view_campaign).collect::<Html>()
+                if let Some(campaign_list) = &query.data {
+                    campaign_list.campaigns.iter().map(view_campaign).collect::<Html>()
                 } else {
                     // TODO: Character skeleton while loading
                     html! { <></> }
